Version 1.1
The old file had to be stopped, edited and then rerun to make any changes to the web page content. Now, the web page has been added as a HTML file, that can be modified and will only be served in the next request made. No restarts needed, unless code is changed.
Also, favicon.ico is now parsed. The file can be replaced with any file of your choice.